+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # See https://keepachangelog.com/en/1.0.0/
 
+## [0.0.49] 2021-08-16
+### Added
+- dependency to `wutsi-platform-core`
+- MQueue heroku deployment
+### Changed
+- Add `@CrossOrigin` annotation to controller code generator
+### Removed
+- CORS Filter code generator
+
 ## [0.0.32] 2021-05-07
 ### Added
 - Auto increase version of SDK